+++ examples/src/main/kotlin/org/partiql/examples/PreventJoinVistor.kt
@@ -1,6 +1,5 @@
 package org.partiql.examples
 
-import com.amazon.ion.system.IonSystemBuilder
 import org.partiql.examples.util.Example
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.syntax.PartiQLParserBuilder
@@ -16,8 +15,7 @@ private class InvalidAstException(message: String) : RuntimeException(message)
  */
 
 class PreventJoinVisitorExample(out: PrintStream) : Example(out) {
-    private val ion = IonSystemBuilder.standard().build()
-    private val parser = PartiQLParserBuilder().ionSystem(ion).build()
+    private val parser = PartiQLParserBuilder().build()
 
     private fun hasJoins(sql: String): Boolean = try {
         val ast = parser.parseAstStatement(sql)