+++ partiql-lang/src/main/kotlin/org/partiql/lang/ast/passes/StatementRedactor.kt
@@ -1,6 +1,5 @@
 package org.partiql.lang.ast.passes
 
-import com.amazon.ion.system.IonSystemBuilder
 import com.amazon.ionelement.api.StringElement
 import org.partiql.lang.ast.SourceLocationMeta
 import org.partiql.lang.ast.sourceLocation
@@ -19,8 +18,7 @@ import org.partiql.lang.syntax.PartiQLParserBuilder
  */
 typealias UserDefinedFunctionRedactionLambda = (List<PartiqlAst.Expr>) -> List<PartiqlAst.Expr>
 
-private val ion = IonSystemBuilder.standard().build()
-private val parser = PartiQLParserBuilder().ionSystem(ion).build()
+private val parser = PartiQLParserBuilder().build()
 private const val maskPattern = "***(Redacted)"
 
 const val INVALID_NUM_ARGS = "Invalid number of args in node"