+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/plan/RelConverter.kt
@@ -22,9 +22,9 @@ internal class RelConverter {
      * As of now, the COMMON property of relation operators is under development, so just use empty for now
      */
     private val empty = Common(
-        schema = emptyMap(),
+        typeEnv = emptyList(),
         properties = emptySet(),
-        metas = emptyMap(),
+        metas = emptyMap()
     )
 
     companion object {
@@ -42,7 +42,7 @@ internal class RelConverter {
                         rel = rel,
                         value = RexConverter.convert(projection.value),
                         at = RexConverter.convert(projection.key),
-                        type = null,
+                        type = null
                     )
                 }
                 // SELECT VALUE ... FROM
@@ -50,6 +50,7 @@ internal class RelConverter {
                     Rex.Query.Collection(
                         rel = rel,
                         constructor = RexConverter.convert(projection.value),
+                        type = null
                     )
                 }
                 // SELECT ... FROM
@@ -57,6 +58,7 @@ internal class RelConverter {
                     Rex.Query.Collection(
                         rel = rel,
                         constructor = null,
+                        type = null
                     )
                 }
             }
@@ -134,10 +136,13 @@ internal class RelConverter {
      */
     private fun convertScan(scan: PartiqlAst.FromSource.Scan) = Rel.Scan(
         common = empty,
-        value = RexConverter.convert(scan.expr),
+        value = when (val expr = scan.expr) {
+            is PartiqlAst.Expr.Select -> convert(expr)
+            else -> RexConverter.convert(scan.expr)
+        },
         alias = scan.asAlias?.text,
         at = scan.atAlias?.text,
-        by = scan.byAlias?.text,
+        by = scan.byAlias?.text
     )
 
     /**
@@ -148,7 +153,7 @@ internal class RelConverter {
         value = RexConverter.convert(scan.expr),
         alias = scan.asAlias?.text,
         at = scan.atAlias?.text,
-        by = scan.byAlias?.text,
+        by = scan.byAlias?.text
     )
 
     /**
@@ -159,7 +164,7 @@ internal class RelConverter {
         else -> Rel.Filter(
             common = empty,
             input = input,
-            condition = RexConverter.convert(expr),
+            condition = RexConverter.convert(expr)
         )
     }
 
@@ -175,7 +180,6 @@ internal class RelConverter {
         select: PartiqlAst.Expr.Select,
         groupBy: PartiqlAst.GroupBy?
     ): Pair<PartiqlAst.Expr.Select, Rel> {
-
         // Rewrite and extract all aggregations in the SELECT clause
         val (sel, aggregations) = AggregationTransform.apply(select)
 
@@ -209,7 +213,7 @@ internal class RelConverter {
             input = input,
             calls = calls,
             groups = groups,
-            strategy = strategy,
+            strategy = strategy
         )
 
         return Pair(sel, rel)
@@ -220,7 +224,7 @@ internal class RelConverter {
      */
     private fun convertGroupByKey(groupKey: PartiqlAst.GroupKey) = binding(
         name = groupKey.asAlias?.text ?: error("not normalized, group key $groupKey missing unique name"),
-        expr = groupKey.expr,
+        expr = groupKey.expr
     )
 
     /**
@@ -321,7 +325,7 @@ internal class RelConverter {
         val fields = from.bindings().map { n ->
             Field(
                 name = Rex.Lit(ionString(n), StaticType.STRING),
-                value = Rex.Id(n, Case.SENSITIVE, Rex.Id.Qualifier.UNQUALIFIED, type = StaticType.STRUCT),
+                value = Rex.Id(n, Case.SENSITIVE, Rex.Id.Qualifier.UNQUALIFIED, type = StaticType.STRUCT)
             )
         }
         return Binding(
@@ -330,7 +334,7 @@ internal class RelConverter {
                 id = "group_as",
                 args = listOf(Rex.Tuple(fields, StaticType.STRUCT)),
                 modifier = Rex.Agg.Modifier.ALL,
-                type = StaticType.STRUCT,
+                type = StaticType.STRUCT
             )
         )
     }
@@ -362,10 +366,23 @@ internal class RelConverter {
      */
     private fun List<PartiqlAst.ProjectItem>.bindings() = map {
         when (it) {
-            is PartiqlAst.ProjectItem.ProjectAll -> binding("*", it.expr)
+            is PartiqlAst.ProjectItem.ProjectAll -> {
+                val path = PartiqlAst.Expr.Path(it.expr, listOf(PartiqlAst.PathStep.PathWildcard()))
+                val bindingName = when (val expr = it.expr) {
+                    is PartiqlAst.Expr.Id -> expr.name.text
+                    is PartiqlAst.Expr.Lit -> {
+                        when (expr.value.type.isText) {
+                            true -> expr.value.stringValue
+                            false -> nextBindingName()
+                        }
+                    }
+                    else -> nextBindingName()
+                }
+                binding(bindingName, path)
+            }
             is PartiqlAst.ProjectItem.ProjectExpr -> binding(
                 name = it.asAlias?.text ?: error("not normalized"),
-                expr = it.expr,
+                expr = it.expr
             )
         }
     }