+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/PartiqlAstExtensionsTests.kt
@@ -1,6 +1,5 @@
 package org.partiql.lang.eval
 
-import com.amazon.ion.system.IonSystemBuilder
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 import org.partiql.lang.ast.SourceLocationMeta
@@ -9,7 +8,7 @@ import org.partiql.lang.syntax.PartiQLParser
 import org.partiql.lang.util.ArgumentsProviderBase
 
 class PartiqlAstExtensionsTests : EvaluatorTestBase() {
-    private val parser = PartiQLParser(IonSystemBuilder.standard().build())
+    private val parser = PartiQLParser()
 
     data class StartingSourceLocationTestCase(val query: String, val expectedSourceLocationMeta: SourceLocationMeta)