+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/evaluatortestframework/VisitorTransformBaseTestAdapter.kt
@@ -1,7 +1,6 @@
 package org.partiql.lang.eval.evaluatortestframework
 
 import org.partiql.lang.CUSTOM_TEST_TYPES
-import org.partiql.lang.ION
 import org.partiql.lang.eval.EvaluationSession
 import org.partiql.lang.eval.visitors.VisitorTransformBase
 import org.partiql.lang.syntax.PartiQLParser
@@ -20,7 +19,7 @@ class VisitorTransformBaseTestAdapter : EvaluatorTestAdapter {
     }
 
     private fun testVistorTransformBase(tc: EvaluatorTestDefinition) {
-        val parser = PartiQLParser(ION, CUSTOM_TEST_TYPES)
+        val parser = PartiQLParser(CUSTOM_TEST_TYPES)
         val ast = parser.parseAstStatement(tc.query)
 
         val clonedAst = defaultTransformer.transformStatement(ast)